ugly high score table
If no high score file is found it creates a fake one with fate data.

If there is one it reads that in to a variable called score board

Once the game is over if you get the highest score it adds you to the list.

Next Step: Check all scores against the list so it can update against any of the 10.