Player changes
Have removed the draw function from player and bullet

Changed how we handle the bullet interaction so now we create a new bullet whenever you press fire as opposed to a singular bullet in a list.

Cleaned up a load of the old commented out code that was confusing matters.